+++ src/fileSystem/index.ts
@@ -50,8 +50,8 @@ export class GistFileSystemProvider implements FileSystemProvider {
   private _pendingWrites = new Subject<WriteOperation>();
 
   private _onDidChangeFile = new EventEmitter<FileChangeEvent[]>();
-  public readonly onDidChangeFile: Event<FileChangeEvent[]> = this
-    ._onDidChangeFile.event;
+  public readonly onDidChangeFile: Event<FileChangeEvent[]> =
+    this._onDidChangeFile.event;
 
   constructor(private store: Store) {
     this._pendingWrites
@@ -91,7 +91,10 @@ export class GistFileSystemProvider implements FileSystemProvider {
 
   private async isDirectory(uri: Uri): Promise<boolean> {
     const { gistId } = getGistDetailsFromUri(uri);
-    let gist = this.store.gists.find((gist) => gist.id === gistId);
+    let gist = this.store.gists
+      .concat(this.store.archivedGists)
+      .concat(this.store.starredGists)
+      .find((gist) => gist.id === gistId);
 
     if (!gist) {
       gist = await getGist(gistId);
@@ -106,6 +109,7 @@ export class GistFileSystemProvider implements FileSystemProvider {
   private async getGistFromUri(uri: Uri): Promise<Gist> {
     const { gistId } = getGistDetailsFromUri(uri);
     let gist = this.store.gists
+      .concat(this.store.archivedGists)
       .concat(this.store.starredGists)
       .find((gist) => gist.id === gistId);
 
@@ -200,6 +204,7 @@ export class GistFileSystemProvider implements FileSystemProvider {
 
       let gist = this.store.gists
         .concat(this.store.starredGists)
+        .concat(this.store.archivedGists)
         .find((gist) => gist.id === gistId);
 
       if (!gist) {
@@ -237,6 +242,7 @@ export class GistFileSystemProvider implements FileSystemProvider {
 
       let gist = this.store.gists
         .concat(this.store.starredGists)
+        .concat(this.store.archivedGists)
         .find((gist) => gist.id === gistId);
 
       if (!gist) {