+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.8.0 (03/18/25)
 
-- Introduced the `GistPad > Sync on Save` setting, which allows you to manually control when edits are synced by to your gist (using the new `GistPad: Sync gist` command). Thanks @junqi-lu for this contribution ❤️
+- Introduced support for archiving gists, which will then display them in a new `Archived gists` node at the bottom of the `Gists` tree view
 - Updated the daily notes gist to be called `📆 Daily notes` in order to align with [GistPad.dev](https://gistpad.dev) and [Gistpad-MCP](https://npmjs.com/gistpad-mcp).
 
 ## v0.7.0 (02/10/25)